Move ExportsHistory into components folder
